Fix eleventh-round review findings
Ten units contradicted an assertion in their own doc comment, templ copy, or the
module doc. No runtime defects — but commit 81c84cf set the standard that this
class blocks on this branch, and applying it only when convenient would make it
worthless.

Three artifacts described the interval-edit behaviour deleted as a bug
-----------------------------------------------------------------------
RescheduledDue shifts by the difference between intervals and leaves owed work
alone. But the plan page told staff the change was "measured from when each was
last done" — that is the max(completed_on) anchor removed in 4bc3f23 precisely
because it could book a visit the customer had declined. The module doc said
"reschedules every pending occurrence", contradicted by its own Rescheduling
section two hundred lines later. And the audit recorded "rescheduled: N" where N
was every occurrence examined: forty overdue machines logged forty moves having
moved none.

"Filter, not sort" was fixed in two places out of four
-------------------------------------------------------
Round ten corrected the create form's help text. The domain comment and the
comment in migration 081 — the one that lives in the schema — still called the
category advisory. It has always been a filter.

Two more sweep promises that early returns broke
-------------------------------------------------
"One audit row a day, whatever happened" sat below an error return, so the
did-it-run signal went dark on exactly the days somebody would look for it —
the same shape as the gauge bug fixed nine lines above it. And "publishes the
gauges, always" still skipped the module-off path, pinning four series at
whatever the shop last had. Off now publishes zeros.

Also
----
The docs claimed a FULL OUTER JOIN drops one-sided rows. It is the join that
drops neither; an inner join is the failure described. The claim appeared twice,
once in the store comment it came from. Corrected in both — a true sentence had
been turned false in the paragraph explaining why the UNION was chosen.

The review count said six. It is eleven, and five of the defects were fixes an
earlier commit claimed and had not made, not three. That paragraph exists to be
accurate about overclaiming.

Ended assignments now drop off the pending scopes. Preserving a booked
occurrence stopped its ticket being orphaned; it should not also keep the row
demanding attention forever for an arrangement that ended.

Typed dates are bounded to ten years either side. A slipped year in a plain
date input anchored the next occurrence from it and took the machine off the due
list for a decade, silently. The cost sums cast minutes to bigint: int32
overflows near 358k minutes and that sum spans every account, so one long-lived
machine would have taken down the whole cross-account report rather than its own
row.

One finding I did not act on
-----------------------------
RepriceTimeEntry accepts an explicit 0, which the review read as re-introducing
a second spelling of "unset". It is not. At settings level a zero labour rate is
refused because Set() reads it as unset; per entry, zero means the shop absorbed
that hour — which is exactly what the settings page offers for travel ("set it
to 0.00 if you absorb the drive") and what rateFor stamps every day. Refusing it
would leave the manual path unable to express a state the automatic one creates.
The distinction was clearly not obvious from the code, so it now carries a
comment and a test that pins both halves.

+++ internal/store/service_tickets.go
@@ -628,7 +628,7 @@ func (s *ServiceTicketStore) Totals(ctx context.Context, tx pgx.Tx, ticketID uui
 		`SELECT COALESCE(SUM(minutes) FILTER (WHERE kind = 'labor'), 0),
 		        COALESCE(SUM(minutes) FILTER (WHERE kind = 'travel'), 0),
 		        COALESCE(SUM(minutes) FILTER (WHERE billable), 0),
-		        COALESCE(SUM(round(minutes * rate_cents / 60.0))
+		        COALESCE(SUM(round(minutes::bigint * rate_cents / 60.0))
 		                 FILTER (WHERE rate_cents IS NOT NULL), 0),
 		        COALESCE(SUM(minutes) FILTER (WHERE rate_cents IS NULL), 0)
 		 FROM service_time_entries WHERE ticket_id = $1`,
@@ -761,7 +761,7 @@ func (s *ServiceTicketStore) CostSummary(ctx context.Context, tx pgx.Tx, f Servi
 	    (SELECT COALESCE(SUM(minutes) FILTER (WHERE billable), 0) FROM entries),
 	    -- Each hour at the rate it was booked at, rounded per entry so the
 	    -- figure matches the per-entry ones a staffer can read off the ticket.
-	    (SELECT COALESCE(SUM(round(minutes * rate_cents / 60.0)), 0) FROM entries
+	    (SELECT COALESCE(SUM(round(minutes::bigint * rate_cents / 60.0)), 0) FROM entries
 	      WHERE rate_cents IS NOT NULL),
 	    (SELECT COALESCE(SUM(minutes) FILTER (WHERE rate_cents IS NULL), 0) FROM entries),
 	    (SELECT count(DISTINCT ticket_id) FROM (
@@ -811,9 +811,10 @@ func (s *ServiceTicketStore) AnyCostedTime(ctx context.Context, tx pgx.Tx) (bool
 // small — it is bounded by "customers we did service work for this quarter".
 //
 // Parts and time entries are normalised into one shape and aggregated together
-// rather than joined as two per-customer aggregates. A FULL OUTER JOIN between
-// them would be the obvious way, and it is the way that quietly drops the
-// customer who had parts but no hours logged, or the reverse.
+// rather than joined as two per-customer aggregates. Joining is the obvious
+// way, and an *inner* join is the one that quietly drops the customer who had
+// parts but no hours logged, or the reverse — the UNION keeps both sides
+// without anybody having to remember which join type does that.
 func (s *ServiceTicketStore) CostByCustomer(ctx context.Context, tx pgx.Tx, f ServiceAccountCostFilter) ([]domain.ServiceAccountCost, error) {
 	var args []any
 
@@ -871,7 +872,10 @@ func (s *ServiceTicketStore) CostByCustomer(ctx context.Context, tx pgx.Tx, f Se
 	           -- NULL rate and are excluded by the same filter that excludes a
 	           -- genuinely uncosted hour, which is why the parts side had to be
 	           -- given the column at all.
-	           COALESCE(SUM(round(minutes * rate_cents / 60.0))
+	           -- bigint: minutes * cents overflows int32 near 358k minutes, and
+	           -- this sum spans every account — one long-lived machine would
+	           -- take the whole report down rather than just its own row.
+	           COALESCE(SUM(round(minutes::bigint * rate_cents / 60.0))
 	                    FILTER (WHERE NOT is_part AND rate_cents IS NOT NULL), 0)
 	                                                                    AS labor_cents,
 	           COALESCE(SUM(minutes) FILTER (WHERE NOT is_part AND rate_cents IS NULL), 0)
